backend/requirements.txt
git-filter-repo==2.47.0
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
whitenoise==6.11.0
